Add 'data' option to DataDescription
`DataDescription` now accepts a `data` option as an alternative to `file` or `files`. This option allows to specify static data as a dictionary of column name and column value pairs.

For example, the following configuration is transformed to a data frame with two columns ('gender' and 'N') and two rows:

```
{
  'data': {
    'gender': ['Male', 'Female'],
    'N': [253084, 334119]
  }
}
```